ingestion/requirements.txt
duckdb>=0.9.2
pandas>=2.0.0
kagglehub>=0.1.0
python-dotenv>=1.0.0
dbt-duckdb>=1.6.0
pyyaml
